hewitt-nic/django_backend/backend/requirements.txt
asgiref
Django
django-cors-headers
djangorestframework
djangorestframework-simplejwt
PyJWT
pytz
sqlparse
pymysql
python-dotenv
mysqlclient
Pillow
pandas
numpy
mathplot